Push code down into component
Simplifies code
