+++ frontend/src/routes/passages/passages/+page.svelte
@@ -4,9 +4,6 @@
   import { onMount } from 'svelte'
   import {
     PUBLIC_BOOK_CODES_FROM_USFM_ONLY_URL,
-    PUBLIC_CHAPTERS_IN_BOOKS_URL,
-    PUBLIC_NT_SURVEY_RG_PASSAGES_URL,
-    PUBLIC_STET_PASSAGES_URL,
     PUBLIC_TAILWIND_SM_MIN_WIDTH
   } from '$env/static/public'
   import { env } from '$env/dynamic/public'
@@ -16,7 +13,6 @@
   import WizardBasket from '$lib/passages/WizardBasket.svelte'
   import { langCodeAndNameStore } from '$lib/passages/stores/LanguageStore'
   import { passagesStore, addPassageReference } from '$lib/passages/stores/PassagesStore'
-  import type { BibleReference } from './model'
   import CheckIcon from '$lib/CheckIcon.svelte'
 
   // For use by Mobile UI
@@ -38,21 +34,8 @@
     return bookCodesAndNames
   }
 
-  async function getChaptersInBooks(
-    apiRootUrl = env.PUBLIC_BACKEND_API_URL,
-    chaptersInBooksUrl = <string>PUBLIC_CHAPTERS_IN_BOOKS_URL
-  ): Promise<Record<string, number[]>> {
-    const response = await fetch(`${apiRootUrl}${chaptersInBooksUrl}`)
-    const chaptersInBooks: Record<string, number[]> = await response.json()
-    if (!response.ok) {
-      console.error(response.statusText)
-      throw new Error(response.statusText)
-    }
-    return chaptersInBooks
-  }
 
   let bookCodesAndNames: Array<[string, string]> = []
-  let chapters: Record<string, number[]> = {}
 
   onMount(() => {
     let langCode = $langCodeAndNameStore.split(',')[0]
@@ -64,11 +47,6 @@
       })
       .catch((err) => console.error(err))
 
-    getChaptersInBooks()
-      .then((chaptersInBooks_) => {
-        chapters = { ...chaptersInBooks_ } // Ensure reactivity with {...blah}
-      })
-      .catch((err) => console.error(err))
   })
 
   function removePassage(id: number) {
@@ -83,116 +61,12 @@
     }
   }
 
-  let selectedBookCode: string = ''
-  let selectedChapter: string = ''
-  let verseReference: string = ''
-  let chaptersForSelectedBook: number[] = []
 
-  const handleBookChange = (event: Event) => {
-    const target = event.target as HTMLSelectElement
-    selectedBookCode = target.value
-    console.log('Book Selected:', selectedBookCode)
-    // Manually reset chapter only when book changes, avoiding reactivity loop
-    selectedChapter = ''
-    chaptersForSelectedBook = chapters[selectedBookCode] || []
-    console.log('Chapters for selected book:', chaptersForSelectedBook)
-  }
-
-  const handleChapterChange = (event: Event) => {
-    const target = event.target as HTMLSelectElement
-    selectedChapter = target.value
-    console.log('Chapter selected:', selectedChapter)
-  }
-
-  const handleVerseInput = (event: Event) => {
-    const target = event.target as HTMLInputElement
-    verseReference = target.value
-  }
-
-
-  async function getNTSurveyRGPassages(
-    langCode: string,
-    apiRootUrl = env.PUBLIC_BACKEND_API_URL,
-    ntSurveyRgPassagesUrl = <string>PUBLIC_NT_SURVEY_RG_PASSAGES_URL
-  ): Promise<Array<BibleReference>> {
-    const url = `${apiRootUrl}${ntSurveyRgPassagesUrl}/${langCode}`
-    console.log(`url: ${url}`)
-    const response = await fetch(url)
-    const bibleReferences: Array<BibleReference> = await response.json()
-    if (!response.ok) {
-      console.error(response.statusText)
-      throw new Error(response.statusText)
-    }
-    return bibleReferences
-  }
-
-  export async function addNTSurveyRGPassages() {
-    try {
-      const bibleReferences = await getNTSurveyRGPassages($langCodeAndNameStore.split(',')[0])
-      console.log(`bibleReferences[0]: ${bibleReferences[0]}`)
-      for (const bibleRef of bibleReferences) {
-          addPassageReference(
-            $langCodeAndNameStore.split(",")[0],
-            bibleRef.book_code,
-            bibleRef.book_name,
-            Number(bibleRef.start_chapter),
-            bibleRef.start_chapter_verse_ref,
-            Number(bibleRef.end_chapter),
-            bibleRef.end_chapter_verse_ref
-          )
-      }
-    } catch (error) {
-      console.error("Failed to add NT Survey RG passages:", error)
-    } finally {
-      console.log("Passages added successfully")
-    }
-  }
-
-  async function getSTETPassages(
-    langCode: string,
-    apiRootUrl = env.PUBLIC_BACKEND_API_URL,
-    stetPassagesUrl = <string>PUBLIC_STET_PASSAGES_URL
-  ): Promise<Array<BibleReference>> {
-    const url = `${apiRootUrl}${stetPassagesUrl}/${langCode}`
-    console.log(`url: ${url}`)
-    const response = await fetch(url)
-    const bibleReferences: Array<BibleReference> = await response.json()
-    if (!response.ok) {
-      console.error(response.statusText)
-      throw new Error(response.statusText)
-    }
-    return bibleReferences
-  }
-
-  export async function addSTETPassages() {
-    try {
-      const bibleReferences = await getSTETPassages($langCodeAndNameStore.split(',')[0])
-      console.log(`bibleReferences[0]: ${bibleReferences[0]}`)
-      for (const bibleRef of bibleReferences) {
-        addPassageReference(
-          $langCodeAndNameStore.split(',')[0],
-          bibleRef.book_code,
-          bibleRef.book_name,
-          Number(bibleRef.start_chapter),
-          bibleRef.start_chapter_verse_ref,
-          Number(bibleRef.end_chapter),
-          bibleRef.end_chapter_verse_ref
-        )
-      }
-    } catch (error) {
-      console.error('Failed to add STET passages:', error)
-    } finally {
-      console.log('Passages added successfully')
-    }
-  }
 
   let windowWidth: number = typeof window !== 'undefined' ? window.innerWidth : 0
   let TAILWIND_SM_MIN_WIDTH: number = PUBLIC_TAILWIND_SM_MIN_WIDTH as unknown as number
 
   $: console.log(`windowWidth: ${windowWidth}`)
-  $: console.log(`selectedBookCode: ${selectedBookCode}`)
-  $: console.log(`selectedChapter: ${selectedChapter}`)
-  $: console.log(`verseReference: ${verseReference}`)
   $: console.log(`$passagesStore: ${JSON.stringify($passagesStore)}`)
 </script>
 
@@ -211,18 +85,7 @@
           />
         </div>
       {:else}
-        <BibleReferenceSelector
-          {bookCodesAndNames}
-          bind:selectedBookCode
-          bind:selectedChapter
-          bind:verseReference
-          {chaptersForSelectedBook}
-          {handleBookChange}
-          {handleChapterChange}
-          {handleVerseInput}
-          {addNTSurveyRGPassages}
-          {addSTETPassages}
-        />
+        <BibleReferenceSelector {bookCodesAndNames} />
         {#if windowWidth < TAILWIND_SM_MIN_WIDTH}
           <button class="ml-2" on:click={() => (showWizardBasketModal = true)}>
             <div class="relative">
