+++ frontend/src/routes/passages/passages/BibleReferenceSelector.svelte
@@ -1,23 +1,150 @@
 <script lang="ts">
   import { passagesStore, addPassageReference } from '$lib/passages/stores/PassagesStore'
   import { langCodeAndNameStore } from '$lib/passages/stores/LanguageStore'
+  import { onMount } from 'svelte'
+  import {
+    PUBLIC_CHAPTERS_IN_BOOKS_URL,
+    PUBLIC_NT_SURVEY_RG_PASSAGES_URL,
+    PUBLIC_STET_PASSAGES_URL,
+    PUBLIC_TAILWIND_SM_MIN_WIDTH
+  } from '$env/static/public'
+  import { env } from '$env/dynamic/public'
+  import type { BibleReference } from './model'
 
   export let bookCodesAndNames: [string, string][] = []
-  export let selectedBookCode: string = ''
-  export let selectedChapter: string = ''
-  export let verseReference: string = ''
-  export let chaptersForSelectedBook: number[] = []
-
-  export let handleBookChange: (event: Event) => void
-  export let handleChapterChange: (event: Event) => void
-  export let handleVerseInput: (event: Event) => void
-  export let addNTSurveyRGPassages: () => Promise<void>
-  export let addSTETPassages: () => Promise<void>
 
   let loading = false
   let ntSurveySuccessMessage: string = ''
   let stetSuccessMessage: string = ''
   let passageSuccessMessage: string = ''
+
+  let selectedBookCode: string = ''
+  let selectedChapter: string = ''
+  let verseReference: string = ''
+  let chaptersForSelectedBook: number[] = []
+  let chapters: Record<string, number[]> = {}
+
+
+  onMount(() => {
+    getChaptersInBooks()
+      .then((chaptersInBooks_) => {
+        chapters = { ...chaptersInBooks_ } // Ensure reactivity with {...blah}
+      })
+      .catch((err) => console.error(err))
+  })
+
+  async function getChaptersInBooks(
+    apiRootUrl = env.PUBLIC_BACKEND_API_URL,
+    chaptersInBooksUrl = <string>PUBLIC_CHAPTERS_IN_BOOKS_URL
+  ): Promise<Record<string, number[]>> {
+    const response = await fetch(`${apiRootUrl}${chaptersInBooksUrl}`)
+    const chaptersInBooks: Record<string, number[]> = await response.json()
+    if (!response.ok) {
+      console.error(response.statusText)
+      throw new Error(response.statusText)
+    }
+    return chaptersInBooks
+  }
+
+
+  async function getNTSurveyRGPassages(
+    langCode: string,
+    apiRootUrl = env.PUBLIC_BACKEND_API_URL,
+    ntSurveyRgPassagesUrl = <string>PUBLIC_NT_SURVEY_RG_PASSAGES_URL
+  ): Promise<Array<BibleReference>> {
+    const url = `${apiRootUrl}${ntSurveyRgPassagesUrl}/${langCode}`
+    console.log(`url: ${url}`)
+    const response = await fetch(url)
+    const bibleReferences: Array<BibleReference> = await response.json()
+    if (!response.ok) {
+      console.error(response.statusText)
+      throw new Error(response.statusText)
+    }
+    return bibleReferences
+  }
+
+  export async function addNTSurveyRGPassages() {
+    try {
+      const bibleReferences = await getNTSurveyRGPassages($langCodeAndNameStore.split(',')[0])
+      console.log(`bibleReferences[0]: ${bibleReferences[0]}`)
+      for (const bibleRef of bibleReferences) {
+        addPassageReference(
+          $langCodeAndNameStore.split(',')[0],
+          bibleRef.book_code,
+          bibleRef.book_name,
+          Number(bibleRef.start_chapter),
+          bibleRef.start_chapter_verse_ref,
+          Number(bibleRef.end_chapter),
+          bibleRef.end_chapter_verse_ref
+        )
+      }
+    } catch (error) {
+      console.error('Failed to add NT Survey RG passages:', error)
+    } finally {
+      console.log('Passages added successfully')
+    }
+  }
+
+  async function getSTETPassages(
+    langCode: string,
+    apiRootUrl = env.PUBLIC_BACKEND_API_URL,
+    stetPassagesUrl = <string>PUBLIC_STET_PASSAGES_URL
+  ): Promise<Array<BibleReference>> {
+    const url = `${apiRootUrl}${stetPassagesUrl}/${langCode}`
+    console.log(`url: ${url}`)
+    const response = await fetch(url)
+    const bibleReferences: Array<BibleReference> = await response.json()
+    if (!response.ok) {
+      console.error(response.statusText)
+      throw new Error(response.statusText)
+    }
+    return bibleReferences
+  }
+
+  export async function addSTETPassages() {
+    try {
+      const bibleReferences = await getSTETPassages($langCodeAndNameStore.split(',')[0])
+      console.log(`bibleReferences[0]: ${bibleReferences[0]}`)
+      for (const bibleRef of bibleReferences) {
+        addPassageReference(
+          $langCodeAndNameStore.split(',')[0],
+          bibleRef.book_code,
+          bibleRef.book_name,
+          Number(bibleRef.start_chapter),
+          bibleRef.start_chapter_verse_ref,
+          Number(bibleRef.end_chapter),
+          bibleRef.end_chapter_verse_ref
+        )
+      }
+    } catch (error) {
+      console.error('Failed to add STET passages:', error)
+    } finally {
+      console.log('Passages added successfully')
+    }
+  }
+
+
+  const handleBookChange = (event: Event) => {
+    const target = event.target as HTMLSelectElement
+    selectedBookCode = target.value
+    console.log('Book Selected:', selectedBookCode)
+    // Manually reset chapter only when book changes, avoiding reactivity loop
+    selectedChapter = ''
+    chaptersForSelectedBook = chapters[selectedBookCode] || []
+    console.log('Chapters for selected book:', chaptersForSelectedBook)
+  }
+
+  const handleChapterChange = (event: Event) => {
+    const target = event.target as HTMLSelectElement
+    selectedChapter = target.value
+    console.log('Chapter selected:', selectedChapter)
+  }
+
+  const handleVerseInput = (event: Event) => {
+    const target = event.target as HTMLInputElement
+    verseReference = target.value
+  }
+
 
   async function handleAddNTSurveyRGPassagesClick() {
     loading = true
